requirements.txt
Django==4.2.0
opencv-python==4.8.1.78
numpy==1.24.3
qrcode==7.4.2
pillow==10.0.0
requests==2.31.0
python-decouple==3.8
